Make the Adaptive Card designer a component (#2035)
* Componentize toolbar

* Work in progress

* Work in progress (cont)

* Work in progress (cont)

* Make designer a component - complete

* Bug fixes, cleanup

* Fix action buttons not updating when modified in designer
