+++ source/nodejs/adaptivecards-designer/css/default-container.css
@@ -1,0 +1,155 @@
+.card-frame {
+    border: 1px solid #F1F1F1;
+    box-shadow: 0 0 15px -5px rgba(0, 0, 0, 0.4);
+    background: white;
+}
+
+.ac-media-poster {}
+
+.ac-media-poster.empty {
+    height: 200px;
+    background-color: #F2F2F2;
+}
+
+.ac-media-playButton {
+    width: 56px;
+    height: 56px;
+    border: 1px solid #EEEEEE;
+    border-radius: 28px;
+    box-shadow: 0px 0px 10px #EEEEEE;
+    background-color: rgba(255, 255, 255, 0.9);
+    color: black;
+    cursor: pointer;
+}
+
+.ac-media-playButton-arrow {
+    color: black;
+}
+
+.ac-media-playButton:hover {
+    background-color: white;
+}
+
+.ac-image.ac-selectable {
+    cursor: pointer;
+}
+
+a.ac-anchor {
+    text-decoration: none;
+}
+
+a.ac-anchor:link {
+    color: #005A9E;
+}
+
+a.ac-anchor:visited {
+    color: #005A9E;
+}
+
+a.ac-anchor:link:active {
+    color: #004D84;
+}
+
+a.ac-anchor:visited:active {
+    color: #004D84;
+}
+
+.ac-container.ac-selectable, .ac-columnSet.ac-selectable {
+    padding: 0px;
+}
+
+.ac-container.ac-selectable:hover, .ac-columnSet.ac-selectable:hover {
+    background-color: rgba(0, 0, 0, 0.1) !important;
+}
+
+.ac-container.ac-selectable:active, .ac-columnSet.ac-selectable:active {
+    background-color: rgba(0, 0, 0, 0.15) !important;
+}
+
+.ac-pushButton {
+    overflow: hidden;
+    text-overflow: ellipsis;
+    text-align: center;
+    vertical-align: middle;
+    cursor: default;
+    font-family: "Segoe UI", sans-serif;
+    font-size: 14px;
+    font-weight: 600;
+    padding: 4px 10px 5px 10px;
+    -webkit-user-select: none;
+    -moz-user-select: none;
+    -ms-user-select: none;
+    user-select: none;
+    height: 31px;
+    background-color: white;
+    color: #0078D7;
+    border: 1px solid #B2E0FF;
+}
+
+.ac-pushButton:hover {
+    background-color: #0078D7;
+    color: white;
+    border: 1px solid #73C1FD;
+}
+
+.ac-pushButton:active {
+    background-color: #464B93;
+    color: white;
+    border: 1px solid #464B93;
+}
+
+.ac-pushButton.subdued {
+    color: #666666;
+    border: 1px solid #EEEEEE;
+}
+
+.ac-pushButton.subdued:hover {
+    background-color: #0078D7;
+    color: white;
+    border: 1px solid #0078D7;
+}
+
+.ac-pushButton.expanded {
+    background-color: #0078D7;
+    color: white;
+    border: 1px solid #0078D7;
+}
+
+.ac-pushButton.expandable:after {
+    -moz-osx-font-smoothing: grayscale;
+    -webkit-font-smoothing: antialiased;
+    display: inline-block;
+    font-family: 'FabricMDL2Icons';
+    font-size: 10px;
+    font-style: normal;
+    font-weight: 900;
+    margin: auto 0px auto 6px;
+    content: '\E70D';
+}
+
+.ac-pushButton.expandable.expanded:after {
+    content: '\E70E';
+}
+
+.ac-input {
+    font-family: "Segoe UI", sans-serif;
+    font-size: 14px;
+    color: black;
+}
+
+.ac-input.ac-textInput {
+    resize: none;
+}
+
+.ac-input.ac-textInput.ac-multiline {
+    height: 72px;
+}
+
+.ac-input.ac-textInput, .ac-input.ac-numberInput, .ac-input.ac-dateInput, .ac-input.ac-timeInput {
+    height: 31px;
+}
+
+.ac-input.ac-textInput, .ac-input.ac-numberInput, .ac-input.ac-dateInput, .ac-input.ac-timeInput, .ac-input.ac-multichoiceInput {
+    border: 1px solid #DDDDDD;
+    padding: 4px 8px 4px 8px;
+}
