+++ source/nodejs/adaptivecards-designer/src/containers/default-container.ts
@@ -1,0 +1,8 @@
+import { HostContainer } from "./host-container";
+
+export class DefaultContainer extends HostContainer {
+    public renderTo(hostElement: HTMLElement) {
+        hostElement.classList.add("card-frame");
+        hostElement.appendChild(this.cardHost);
+    }
+}
