+++ source/nodejs/adaptivecards-designer/src/fullscreen-handler.ts
